+++ src/Components/Database/ConnectionFactory.php
@@ -2,6 +2,7 @@
 
 namespace Craft\Components\Database;
 
+use Craft\Components\Database\MySql\Connection;
 use Craft\Contracts\ConnectionFactoryInterface;
 use Craft\Contracts\DataBaseConnectionInterface;
 
@@ -14,7 +15,7 @@ class ConnectionFactory implements ConnectionFactoryInterface
     public function createConnection(array $config): DataBaseConnectionInterface
     {
         return match ($config['driver']) {
-            'mysql' => new QueryBuilder($config),
+            'mysql' => new Connection($config),
         };
     }
 }