+++ src/Console/Events.php
@@ -2,32 +2,12 @@
 
 namespace Craft\Console;
 
-class Events
+enum Events: string
 {
-    /**
-     * @param string $message
-     */
-    const BEFORE_EXECUTE = 'before_execute';
-
-    /**
-     * @param string $message
-     */
-    const AFTER_EXECUTE = 'after_execute';
-
-    /**
-     * @param string $message
-     */
-    const BEFORE_RUN = 'before_run';
-
-    /**
-     * @param string $message
-     */
-    const OPTIONS_CONFIRM = 'options_confirm';
-
-    /**
-     * @param string $message
-     */
-    const OPTION_CONFIRMED = 'option_confirmed';
-    
-    const FILE_SAVED = 'file_saved';
+    case BEFORE_EXECUTE = 'before_execute';
+    case AFTER_EXECUTE = 'after_execute';
+    case BEFORE_RUN = 'before_run';
+    case OPTIONS_CONFIRM = 'options_confirm';
+    case OPTION_CONFIRMED = 'option_confirmed';
+    case FILE_SAVED = 'file_saved';
 }