+++ src/Http/Message/Request.php
@@ -8,7 +8,7 @@ use Craft\Contracts\UriInterface;
 use Craft\Http\Processors\AuthHeaderProcessor;
 use Craft\Http\Processors\FormDataProcessor;
 
-class Request extends Message implements RequestInterface
+class Request implements RequestInterface
 {
     /**
      * @var string
@@ -177,7 +177,13 @@ class Request extends Message implements RequestInterface
      */
     public function getHeaderLine(string $name): string
     {
-        return implode(',', $this->headers[$name] ?? []);
+        $data = $this->headers[$name] ?? [];
+
+        if (is_string($data) === true) {
+            return $data;
+        }
+
+        return implode(',', $data ?? []);
     }
 
     /**